+++ src/letter-sidc/icons-signalsIntelligence.js
@@ -438,7 +438,7 @@ module.exports = function signalsIntelligence(sId, bbox, icn, _STD2525) {
   sId["I-U-SRE---"] = [
     icn["SI.IC.RADAR"],
     icn["SI.M1.ECHO"],
-    icn["SI.M2.WHISKY"]
+    icn["SI.M2.WHISKEY"]
   ];
   sId["I-U-SRM---"] = [
     icn["SI.IC.RADAR"],